final lint fixes
