+++ farmer/__init__.py
@@ -1,8 +1,2 @@
-import farmer.ext.output as output
-import farmer.ext.prompt as prompt
-from farmer.ext.area import Area
-from farmer.ext.asset import Animal, Asset, Planting
-from farmer.ext.farm import Farm, farm
-from farmer.ext.log import Harvest, Log, Observation, Seeding, Transplanting
-from farmer.ext.others import Content, Measure, Quantity, User
-from farmer.ext.term import Category, Crop, CropFamily, Season, Term, Unit
+
+from farmer.ext import Farm
