+++ farmer/ext/__init__.py
@@ -1,1 +1,10 @@
 """Extension imports."""
+
+import farmer.ext.output as output
+import farmer.ext.prompt as prompt
+from farmer.ext.area import Area
+from farmer.ext.asset import Animal, Asset, Planting, Equipment
+from farmer.ext.farm import Farm
+from farmer.ext.log import Harvest, Log, Observation, Seeding, Transplanting
+from farmer.ext.others import Content, Measure, Quantity, User
+from farmer.ext.term import Category, Crop, CropFamily, Season, Term, Unit
