+++ farmer/reporting/__init__.py
@@ -1,2 +1,3 @@
 
+from farmer.reporting.report import Report
 from farmer.reporting.rstreport import RstReporter
